+++ android/app/src/main/kotlin/deckers/thibault/aves/ScreenSaverSettingsActivity.kt
@@ -1,9 +1,10 @@
 package deckers.thibault.aves
 
 import android.content.Intent
+import deckers.thibault.aves.model.FieldMap
 
 class ScreenSaverSettingsActivity : MainActivity() {
-    override fun extractIntentData(intent: Intent?): MutableMap<String, Any?> {
+    override fun extractIntentData(intent: Intent?): FieldMap {
         return hashMapOf(
             INTENT_DATA_KEY_ACTION to INTENT_ACTION_SCREEN_SAVER_SETTINGS,
         )