+++ lib/utils/xmp_utils.dart
@@ -1,3 +1,4 @@
+import 'package:aves/ref/locales.dart';
 import 'package:aves/ref/metadata/xmp.dart';
 import 'package:intl/intl.dart';
 import 'package:xml/xml.dart';
@@ -60,7 +61,7 @@ class XMP {
     return '${offsetMinutes.isNegative ? '-' : '+'}${h.toString().padLeft(2, '0')}:${m.toString().padLeft(2, '0')}';
   }
 
-  static String toXmpDate(DateTime date) => '${DateFormat('yyyy-MM-ddTHH:mm:ss').format(date)}${_xmpTimeZoneDesignator(date)}';
+  static String toXmpDate(DateTime date) => '${DateFormat('yyyy-MM-ddTHH:mm:ss', asciiLocale).format(date)}${_xmpTimeZoneDesignator(date)}';
 
   static String? getString(
     List<XmlNode> nodes,