+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - opening app from launcher always show home page
+- use dates with western arabic numerals for maghreb arabic locales
 - upgraded Flutter to stable v3.19.4
 
 ### Fixed